requirements.txt
fastapi
uvicorn[standard]
gradio
python-dotenv
chromadb==0.5.5
sentence-transformers
transformers
accelerate
requests
